+++ src/pages/Discord/Discord.Code.tsx
@@ -1,6 +1,5 @@
 import type React from "react";
-import CopyToClipboardButton from "../../components/ui/custom/copy-to-clipboard-button";
-import RefreshButton from "../../components/ui/custom/refresh-button";
+import { RefreshCcw, CopyIcon } from "lucide-react";
 
 interface AuthCodeProps {
   code: string;
@@ -9,12 +8,34 @@ interface AuthCodeProps {
 }
 
 const DiscordCode: React.FC<AuthCodeProps> = ({ code, onRefresh }) => {
+  const handleCopy = () => {
+    navigator.clipboard.writeText(code);
+  };  
   return (
     <div className="flex h-36 flex-col items-center gap-4">
-      <p className="font-medium text-5xl">{code}</p>
+      <p 
+        className="font-medium text-5xl cursor-pointer p-4 rounded-4xl transition-colors hover:bg-neutral-200"
+        onClick={onRefresh}
+      >
+        {code}
+      </p>
       <div className="flex gap-4">
-        <RefreshButton onClick={onRefresh} />
-        <CopyToClipboardButton textToCopy={code} />
+      <button
+        onClick={onRefresh}
+        type="button"
+        className="flex items-center gap-1 text-neutral-500 underline cursor-pointer p-1 rounded-2xl transition-colors hover:bg-neutral-200"
+      >
+        <RefreshCcw size="12" />
+        <span>새로고침</span>
+      </button>
+      <button
+        onClick={handleCopy}
+        type="button"
+        className="flex items-center gap-1 text-slate-500 underline cursor-pointer p-1 rounded-2xl transition-colors hover:bg-neutral-200"
+      >
+        <CopyIcon size={16}/>
+        <span>복사하기</span>
+      </button>
       </div>
     </div>
   );
